refactor persistent storage (#170)
HelloFresh does actually need experiment name to solve for concurrent
writes across FE clients for the same user but different experiments.
For example, userA persists experiment1 and experiment2 in FE1 and
experiment3 in FE2. The storage needs to accept all 3 but if `Save`
receives them bundled as (experiment1, experiment2) and (experiment3),
it's difficult to handle the merging.

Client "spec freshness" time is necessary for multiple evaluation of the
same experiment.
For example, userA persists experiment1 in FE1 and experiment1 in FE2.
The storage needs to accept the latest and reject the oldest.

Since we are no longer following the model of "the user doesn't need to
care what the persisted values looks like",
we should also no longer handle the JSON-ificiation. Instead, we pass
the entire `StickyValues` object and expect the interface to `Load` a
map of experiments to `StickyValues` for a user

https://github.com/statsig-io/kong/pull/1914