Add Transport Option to allow egress proxies
This commit adds a `Transport` to the `Options` struct when constructing
a new instance of the Go SDK. If set it allows the end user to pass in
their own `http.RoundTripper` instead of using `http.DefaultTransport`.

The reasoning for this change it to allow users to provide a transport
which is capabile of acting as middleware for the outbound requests made
by the SDK (such as allowing routing via an authenticated egress proxy).