fix race condition in init context (#238)
There is still a race condition on the init context. This occurs only in the event of a timeout. The goroutine actually continues executing, except we just unblock the main process (similar to js promise.race). So there is a race condition when writing to `context.Error` (1. in the timeout handler, 2. during `client.init`)
To solve, use a mutex lock and clone to context. (Cloning is necessary because even if we use a mutex lock, there is no guarantee that "Timed out" will be the final error set)