fix: fallback to statsig on network err (#320)
customer thread here:
https://statsigconnect.slack.com/archives/C084S8HH410/p1763103687994729?thread_ts=1762960138.782269&cid=C084S8HH410

go early return on errors so it doesn't actually retry and fallback to
statsig. it also doesn't fallback to statsig on any non retryable status
code. even tho in both cases the sdk allows for extra retry count, it's
returned as non retryable request. also copy headers onto retry request.

also fixing a bug introduced during the additional of this fallback
logic. because we preconstruct the retry request, we are setting the new
url as the source api even though it may not have actually been used.
this returns the incorrect url in initializeDetails and logging.
correcting it to only set the source api when we fully parsed the
response.